pyproject.toml
[project]
name = "receipt-lens"
version = "0.1.0"
description = "Drop in a photo of a receipt, get back structured data. Runs locally with vision models."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Kamil Palubicki", email = "palubicki.kamil@gmail.com" }
]
keywords = ["receipt", "ocr", "vision", "ollama", "local-ai", "finance"]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-multipart>=0.0.19",
    "pillow>=11.0",
    "httpx>=0.27",
    "ollama>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["receipt_lens"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    "integration: marks tests that require Ollama with a vision model",
]
